requirements.txt
face-recognition
opencv-python
numpy
Flask
cloudinary
python-dotenv
requests
